src/containers/http-server/requirements.txt
celery
flask
requests
biopython
